+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased
+## 10.0.0 - *2021-01-25*
 
 - Add `mysql_database` resource for creating, querying and removing databases
 - Add `mysql_user` resource for creating, modifying and removing database users